[refactor] move Game class to its own module
